wip: better docs examples
