+++ gatsby-theme-woly/package.json
@@ -41,10 +41,11 @@
     "gatsby-plugin-styled-components": "^3.10.0",
     "gatsby-source-filesystem": "^2.11.0",
     "gatsby-transformer-sharp": "^3.1.0",
+    "parse-imports": "^1.1.0",
     "prism-react-renderer": "^1.1.1",
     "react-helmet": "^6.1.0",
     "remark-comments": "^1.2.9",
-    "unist-util-visit": "^3.1.0"
+    "rimraf": "^3.0.2"
   },
   "devDependencies": {
     "babel-plugin-styled-components": "^1.12.0",
